Corrections Warnings VS 2005 Pro (Debug & Release)

+++ scilab/routines/graphics/Xcall1.c
@@ -1009,7 +1009,7 @@ void xstringb_bbox (char *string, integer x, integer y, integer w, integer h, do
     }
     
     x1=x+ (w- wmax)/2;
-    yy1=y - h + ( h - htot)/2 + rect[3];
+    yy1=(integer)(y - h + ( h - htot)/2 + rect[3]);
     
     bbox[0] = x1;
     bbox[1] = yy1 - hauteur;
